Slightly better subscripts and superscripts in textuals
CSchem textuals now render "f_0" as "f₀" and "x^2" as "x²".
Implementation is slightly crude: the "_" and "^" are not fully
suppressed in the html output, but rather enclosed in a span that
makes them hard to see: font-size=2pt and color=white.
The SVG exporter does not yet support this.