Reverted Alt scrolling in CSchem
Bizarrely, on my home computer, a physical press+release of either
Alt key results in a single KeyPressEvent. A second press+release
then results in a single KeyReleaseEvent. As a consequence, a user
action like activating the menu (e.g. Alt-F) causes the program
to stay stuck in "hand drag" mode.

There has to be a better way.