Bugfix to CPCB: Zero-length traces
It was possible to create traces of zero length in CPCB by taking
a legit trace and dragging its end onto its beginning.
Zero-length traces were not drawn on screen. (Qt peculiarity?)
However, they did get drawn into gerber output. This created an
incongruence between screen and gerber.
This commit:
1. Quietly drops zero-length traces upon loading;
2. Deletes zero-length traces when they are created;
3. Prevents the gerber write from writing zero-length traces, so
   as to restore parity between screen and gerber.
In the long run, (3) may not be desirable, because IMHO, the gerber
behavior is more reasonable than the Qt (?) behavior on zero-length
lines with round caps.